agents_catalog/37-Agentic-Compliance-Lab/requirements.txt
aws-cdk-lib>=2.150.0
aws-cdk.aws-bedrock-agentcore-alpha
constructs>=10.0.0
